Clean up; deduplicate function.
